refactor: extract fake zeebe execution listener
This class was annoyingly in the middle of the test class, breaking up
the ability to read test cases.

We looked into replacing it with Mockito, but that required more lines
per test case, and was significantly slower (on test initialization).

(cherry picked from commit 7c431ceb5299a6a1b75831bab27a591db822a7c3)
